requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
blessings==1.7
bokeh==2.3.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dateparser==1.0.0
et-xmlfile==1.1.0
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-api-core==1.31.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gpustat==0.6.0
greenlet==1.1.1
grpcio==1.34.1
h5py==3.1.0
hiredis==2.0.0
idna==3.2
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
msgpack==1.0.2
multidict==5.1.0
numba==0.53.1
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus==0.7.13
opencensus-context==0.1.2
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.1
Pillow==8.3.1
prometheus-client==0.11.0
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
py-spy==0.3.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pydantic==1.8.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.22
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.6.3
